test: Add a snapshot test with Jest
Added the Jest package.
Added a simple snapshot test for the Article component.
Configured ESLint to recognize the global variables introduced by Jest.
Configured Babel to transpile ES modules to commonJS modules only in the `test` environment (Jest runs in Node and it requires transpilation for this feature - see docs `https://jestjs.io/docs/en/webpack#using-with-webpack-2`).
Configured Babel to not transpile ES modules (apart from the test environment), as it is a job that will be performed by Webpack and it would also prevent tree-shaking by Webpack.