refactor: Refactor Login, create Form and add test
Refactored the Login component by creating a Form subcomponent that could then be called as its child.
Refactored the Login `onSignupSubmit` method to be called by this child component as a function with the submitted user as argument. I got the idea from a workshop about testing components on Frontend Masters (see the `Form` test file for reference).
Added a unit test for the newly created Form component.